+++ fizzbuzz/benches/bench_fizzbuzz.rs
@@ -1,5 +1,5 @@
 use criterion::{criterion_group, criterion_main, Criterion};
-use fizzbuzz::{self, FizzBuzz};
+use fizzbuzz::{self, FizzBuzz, MultiFizzBuzz};
 
 #[inline]
 fn fb100_000() {
@@ -10,8 +10,16 @@ fn fb100_000() {
     }
 }
 
+#[inline]
+fn fb100_000_vec() {
+    for _ in 1..10 {
+        (1..100_000).collect::<Vec<_>>().fizzbuzz();
+    }
+}
+
 fn criterion_benchmark(c: &mut Criterion) {
     c.bench_function("fizzbuzz 100_000", |b| b.iter(|| fb100_000()));
+    c.bench_function("fizzbuzz 100_000_vector", |b| b.iter(|| fb100_000_vec()));
 }
 
 criterion_group!(benches, criterion_benchmark);